requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
python-dotenv
itsdangerous
pandas
openpyxl
pymongo
